CR-272 Refactor CRUD operations to repository package
