+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/repository/RespondentDataRepository.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.integration.rhsvc.service;
+package uk.gov.ons.ctp.integration.rhsvc.repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,7 +6,8 @@ import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.common.event.model.CollectionCase;
 import uk.gov.ons.ctp.common.event.model.UAC;
 
-public interface RespondentDataService {
+/** Repository for Respondent Data */
+public interface RespondentDataRepository {
 
   void writeUAC(UAC uac) throws CTPException;
